requirements.txt
openai
pyyaml
streamlit